requirements.txt
schedule==1.2.0
scrapy>=2.5.0
Pillow>=8.1.1
requests>=2.25.0
lxml>=4.6.0
pymongo>=3.12.0
pytesseract>=0.3.8
paddleocr>=2.7.0
paddlepaddle>=2.5.0
